Refactors pages and api into single files
