+++ public/static/locales/en/form-builder.json
@@ -187,12 +187,13 @@
   "viewCode": "View code",
   "contactus": {
     "title": "Contact Us",
-    "paragraph1": "If you experience technical issues, please fill out the",
-    "paragraph1Link": "Support Form",
-    "paragraph2": "For other enquiries, go to",
-    "paragraph2Link": "Canada.ca Contact us",
-    "paragraph2Part2": "to find the department or agency that can support you.",
-    "paragraph3": "The GC Forms team provides support to people who work in the government. We cannot give general advice to the public about the forms that we host.",
+    "useThisForm": "Use this form to learn more about GC Forms and how it can work in your department.",
+    "gcFormsTeamPart1": "The GC Forms team provides services to people who work in the government. We cannot give general advice to the public about the forms that we host. For other enquiries, go to",
+    "gcFormsTeamLink": "Canada.ca Contact",
+    "gcFormsTeamPart2": "us to find the department or agency that can help you.",
+    "needSupport": "Need Support?",
+    "ifYouExperience": "If you experience technical issues, please fill out the",
+    "supportFormLink": "Support Form",
     "name": "Your Name",
     "email": "Your Email",
     "request": {
@@ -208,14 +209,14 @@
     }
   },
   "support": {
-    "title": "Support",
-    "paragraph1": "If you would like to learn more about the product or set up a demo, please fill out the",
-    "paragraph1Link": "Contact us form",
-    "backToForms": "Back to GC Forms",
-    "paragraph2": "For other enquiries, go to",
-    "paragraph2Link": "Canada.ca Contact us",
-    "paragraph2Part2": "to find the department or agency that can support you.",
-    "paragraph3": "The GC Forms team provides support to people who work in the government. We cannot give general advice to the public about the forms that we host.",
+    "title": "Get support",
+    "useThisForm": "Use this form if you experience technical issues with GC Forms.",
+    "gcFormsTeamPart1": "The GC Forms team provides services to people who work in the government. We cannot give general advice to the public about the forms that we host. For other enquiries, go to",
+    "gcFormsTeamLink": "Canada.ca Contact us",
+    "gcFormsTeamPart2": "to find the department or agency that can help you.",
+    "lookingForADemo": "Looking for a product demo?",
+    "ifYouWouldLike": "If you would like to learn more about GC Forms or set up a demo, please fill out the ",
+    "contactUs": "Contact us form",
     "name": "Your Name",
     "email": "Your Email",
     "request": {
@@ -236,7 +237,8 @@
     "paragraph2Part1": "For other enquiries, go to",
     "paragraph2Link": "Canada.ca Contact us",
     "paragraph2Part2": "to find the department or agency that can support you.",
-    "paragraph3": "The GC Forms team provides support to people who work in the government. We cannot give general advice to the public about the forms that we host."      
+    "paragraph3": "The GC Forms team provides support to people who work in the government. We cannot give general advice to the public about the forms that we host.",
+    "backToForms": "Back To Forms"
   },
   "mainNavAriaLabel": "Main",
   "defaultElementDescription": {
